requirements.txt
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==1.17.1
channels==4.2.2
channels_redis==4.3.0
constantly==23.10.4
cryptography==45.0.5
daphne==4.2.1
Django==5.2.4
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.1.0
redis==6.2.0
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
zope.interface==7.2
